src/main/java/com/siyanda/bankatmweb/repository/BankRepository.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */

package com.siyanda.bankatmweb.repository;

import com.siyanda.bankatmweb.domain.Bank;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.stereotype.Repository;

/**
 *
 * @author inathi-zenande
 */
@Repository
public interface BankRepository  extends JpaRepository<Bank,Long>{

    public void save(Bank bank);

    public Bank findOne(Long id);

    public void delete(Bank bank);

    public void deleteAll();
    
}
